+++ fixture/src/ExamplesScreen.tsx
@@ -18,6 +18,7 @@ export const ExamplesScreen = () => {
   const data = [
     { title: "List", destination: NavigationKeys.LIST },
     { title: "PaginatedList", destination: NavigationKeys.PAGINATED_LIST },
+    { title: "Reminders", destination: NavigationKeys.REMINDERS },
     { title: "Twitter Timeline", destination: NavigationKeys.TWITTER },
     {
       title: "Twitter FlatList Timeline",
@@ -28,6 +29,7 @@ export const ExamplesScreen = () => {
     <>
       <StatusBar barStyle="dark-content" />
       <FlatList
+        testID="ExamplesFlatList"
         keyExtractor={(item) => item.destination}
         data={data}
         renderItem={({ item }) => (